Replace react-toastify with sonner for toasts
Swapped out the react-toastify library with the sonner package across the app to standardize toast notifications. Integrated a new Toaster component that hooks into the theming system for customized styling. This change results in a more cohesive look and feel for the toasts and reduces the styling complexity by leveraging theme-based classNames. Removed react-toastify imports and styles accordingly.